backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.9

# Data validation
pydantic==2.10.3
pydantic-settings==2.6.1

# OpenAI SDK
openai==1.57.0

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.20

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
hypothesis==6.122.3
httpx==0.28.1

# Retry logic
tenacity==9.0.0
